refactor: update import paths for authentication, user, and post actions to improve code organization and maintainability
