+++ components/create-post.tsx
@@ -11,7 +11,7 @@ import UserAvatar from "./user-avatar"
 import { Textarea } from "./ui/textarea"
 import { Button } from "./ui/button"
 import { useActionState, useEffect, useState } from "react";
-import { submitPost } from "@/lib/utils/supabase/actions/post/post.actions";
+import { submitPost } from "@/lib/actions/post/post.actions";
 import { PostSchemaErrorType, PostSchemaType } from "@/lib/validations/postSchema";
 import { usePosts } from "@/lib/context/posts.context";
 import { usePathname } from "next/navigation";
